break up main script into some modules and combine with webpack
