+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 config.json
 .DS_Store
 npm-debug.log
+dist
