+++ package.json
@@ -4,11 +4,13 @@
   "description": "review changes across multiple repositories",
   "main": "app.js",
   "scripts": {
-    "test": "grunt test"
+    "test": "grunt test",
+    "build": "webpack -p",
+    "watch": "webpack -w"
   },
   "repository": {
     "type": "git",
-    "url": "http://github.com/pulseenergy/panorama.git"
+    "url": "http://github.com/wachunga/panorama.git"
   },
   "keywords": [
     "app",
@@ -34,6 +36,7 @@
   "devDependencies": {
     "grunt": "~0.4.1",
     "grunt-contrib-jshint": "~0.10.0",
-    "replay": "^1.12.0"
+    "replay": "^1.12.0",
+    "webpack": "^3.8.1"
   }
 }
