+++ views/include/layout.ejs
@@ -16,7 +16,7 @@
 	<script src="js/underscore.min.js" type="text/javascript"></script>
 	<script src="js/moment.min.js" type="text/javascript"></script>
 	<script src="js/svg.min.js" type="text/javascript"></script>
-	<script src="js/panorama.js" type="text/javascript"></script>
+	<script src="panorama.js" type="text/javascript"></script>
 	<script>
 		var panorama = new Panorama(<%- JSON.stringify(user ? user.organizations : [{ login: 'github' }]) %>);
 		window.onload = function () {
